msm8996: add proxy device to direct pcm and raw outputs
During a switch between WFD and devices that support direct pcm
and raw outputs, the stream gets invalidated which triggers
a teardown event in NuPlayer that causes frame drops during
video playback.

Stream invalidation is not expected as there is no switch between
outputs but as AUDIO_DEVICE_OUT_PROXY is not present in direct pcm
and raw outputs, AudioPolicyManager triggers invalidation of the stream.

Add the device to these outputs to prevent invalidation.

CRs-Fixed: 930869
Change-Id: I94afdd9e4a8805fa4015270a4a8f337971af813f