oneplus3: enable Adaptive QP
This change enables Adaptive QP (PQ),
a feature which is intended to improve
the quality of the encoder by accessing
newly generated GPU statistics and optimizing
the encoder based on the gathered information.

The GPU statistics are generated by "libgpustats"
which was added to the BSP in N.

Since the feature could potentially conflict with OEM
extensions for the encoder, Qualcomm decided to disable it
by default and let the OEMs decide whether they want
to enable it.

We are enabling PQ, since we don't have massive
changes in the handling of the encoder and
no issues were detected when using PQ.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>