oneplus3: Configure Keypress boost
This patch enables and configures
Keypress boost as follows:

The weak boost forces 1.3Ghz on the
little cluster and 1.4Ghz on the big
cluster.

The strong boost forces 1.4Ghz on the little
cluster and 1.7Ghz on the big cluster.
Also the CPUBW frequency is boosted to
the maximum and power collapse is disabled.

The boost duration is determined individually
and automatically, so there is no need to configure it
here.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>